requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
python-terraform==0.10.1
ansible-runner==2.3.4
pyvmomi==8.0.1.0.1
celery[redis]==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
prometheus-client==0.19.0
structlog==24.1.0
pyyaml==6.0.1
jinja2==3.1.3
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
aiofiles==23.2.1
python-multipart==0.0.6
